Move tasks/bulk code
