+++ corehq/apps/commtrack/tasks.py
@@ -1,25 +1,13 @@
 from celery.task import task
 from corehq.apps.commtrack.consumption import recalculate_domain_consumption
 from corehq.apps.locations.bulk import import_locations
-from corehq.apps.commtrack.bulk import import_products
-from dimagi.utils.excel_importer import SingleExcelImporter, MultiExcelImporter
+from dimagi.utils.excel_importer import MultiExcelImporter
 
 
 @task
 def import_locations_async(domain, file_ref_id):
     importer = MultiExcelImporter(import_locations_async, file_ref_id)
     results = list(import_locations(domain, importer))
-    importer.mark_complete()
-
-    return {
-        'messages': results
-    }
-
-
-@task
-def import_products_async(domain, file_ref_id):
-    importer = SingleExcelImporter(import_products_async, file_ref_id)
-    results = import_products(domain, importer)
     importer.mark_complete()
 
     return {
