+++ corehq/apps/products/tasks.py
@@ -1,0 +1,14 @@
+from celery.task import task
+from corehq.apps.products.bulk import import_products
+from dimagi.utils.excel_importer import SingleExcelImporter
+
+
+@task
+def import_products_async(domain, file_ref_id):
+    importer = SingleExcelImporter(import_products_async, file_ref_id)
+    results = import_products(domain, importer)
+    importer.mark_complete()
+
+    return {
+        'messages': results
+    }
